[Application] Move Application classes to new Kernel namespace.
