+++ src/Valkyrja/Application/Kernel/Contract/Application.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Application\Contract;
+namespace Valkyrja\Application\Kernel\Contract;
 
 use Valkyrja\Application\Data\Config;
 use Valkyrja\Application\Data\Data;
