+++ tests/Classes/Dispatch/InvalidDispatcherClass.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Tests\Classes\Dispatch;
 
-use Valkyrja\Application\Contract\Application;
+use Valkyrja\Application\Kernel\Contract\Application;
 
 /**
  * Invalid dispatcher class to test with.
